+++ docs/changes.rst
@@ -4,7 +4,9 @@ Changelog
 11.0.3 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Remove the polyfills-loader.js script. Loading the few polyfills we need in
+  Patternslib is done by Patternslib itself.
+  [thet]
 
 
 11.0.2 (2025-04-01)