+++ docs/changes.rst
@@ -8,6 +8,9 @@ Changelog
   Patternslib is done by Patternslib itself.
   [thet]
 
+- Modify the form survey version to update the tool title and obsolete field
+  [ale-rt]
+
 
 11.0.2 (2025-04-01)
 -------------------